Pipelines/unstructured_pipeline/requirements.txt
# File: requirements.txt

# Core Dependencies
chromadb
neo4j
sentence-transformers
spacy
transformers
torch

# Data Processing
numpy
pandas
tqdm

# Utilities
python-dotenv
pydantic

# # Optional: Advanced Features
# langchain
# openai

# Development
pytest
black
flake8
requests==2.31.0
PyYAML==6.0.1

